Change block header

+++ src/consensus/params.h
@@ -173,7 +173,7 @@ struct Params
     /** Nov 15, 2020 MTP activation time will be 12:00:00 UTC */
     uint64_t nov2020ActivationTime;
     /** Nov 15, 2020 actication height */
-    boost::optional<int> nov2020Height;
+    boost::optional<int64_t> nov2020Height;
     /** May 15, 2021 MTP activation time will be 12:00:00 UTC */
     uint64_t may2021ActivationTime;
 };
